pyproject.toml
[tool.ruff]
line-length = 120
select = ["E", "F", "I"]

[tool.black]
line-length = 120

[project]
name = "yamlp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "fastapi>=0.115.8",
    "ipykernel>=6.29.5",
    "lancedb>=0.21.1",
    "modal>=0.73.67",
    "pillow>=11.1.0",
    "pytest>=8.3.5",
    "python-fasthtml>=0.12.1",
    "requests>=2.32.3",
    "ruff>=0.9.7",
    "sqlmodel>=0.0.24",
]
